Find now accepts a []byte argument.
Assignment to a *[]byte does no copying and can be used as a cheap
way to find the position of a submatch.

Use base 0 in all numeric conversions.  Callers can avoid an unnecessary
hexadecimal parse by changing the regexp to only match [0-9].  It is a
bit harder to avoid an octal parse, but still possible:
   0*([1-9][0-9]*|0)
Or more direct: pass a func that does the appropriate conversion.