Introduce NotFound error.
Callers can look for NotFound using errors.Is to figure out when Scan could not
find the regexp in the input.  Callers using Scan to scan for multiple matches
across a string can single out this error to identify when to stop scanning.

Because we do this by wrapping the NotFound error (and expect the caller to use
errors.Is to match it), this now requires Go 1.13. go.mod has been updated
accordingly.