Support for continuing to scan the same string. (#1)
This allows the caller to iteratively scan across the input until the string is
exhausted.

* Add re.Span, which stores the position of the full match.

* Introduce NotFound error.

Callers can look for NotFound using errors.Is to figure out when Scan could not
find the regexp in the input.  Callers using Scan to scan for multiple matches
across a string can single out this error to identify when to stop scanning.

Because we do this by wrapping the NotFound error (and expect the caller to use
errors.Is to match it), this now requires Go 1.13. go.mod has been updated
accordingly.