re.ScanString matches a regexp against a string.
re.ScanString is a convenience wrapper around re.Scan that converts the
argument from []byte to string, so that the user doesn't have to do so.
Since String functions in package regexp are used much more frequently than the
Bytes functions, removing the fricton of an explicit Bytes conversion from this
package should help adoption.